added delete confirmation windows

+++ client/src/Pages/User/Announcements/ACard/ACard.jsx
@@ -8,41 +8,41 @@ const ACard = ({ id, date, description, groupOrSubgroup }) => {
   const [deleteBtnText, setDeleteBtnText] = useState('Delete');
   const [showFullDescription, setShowFullDescription] = useState(false);
 
-
   useEffect(() => {
     setIsAdmin(userData.role === 'admin');
   }, [userData]);
 
   const handleDelete = async () => {
-    setDeleteBtnText('Deleting...');
-    try {
-      const backendRoute =
-        groupOrSubgroup === 'group'
-          ? `http://localhost:5000/announcement/delGrpAnnouncement/${id}`
-          : `http://localhost:5000/announcement/delSubGrpAnnouncement/${id}`;
+    const confirmed = window.confirm('Are you sure you want to delete this card?');
+    if (confirmed) {
+      setDeleteBtnText('Deleting...');
+      try {
+        const backendRoute =
+          groupOrSubgroup === 'group'
+            ? `http://localhost:5000/announcement/delGrpAnnouncement/${id}`
+            : `http://localhost:5000/announcement/delSubGrpAnnouncement/${id}`;
 
-      const token = localStorage.getItem('jwtoken');
-      const response = await fetch(backendRoute, {
-        method: 'DELETE',
-        credentials: 'include',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}`,
-        },
-      });
+        const token = localStorage.getItem('jwtoken');
+        const response = await fetch(backendRoute, {
+          method: 'DELETE',
+          credentials: 'include',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${token}`,
+          },
+        });
 
-      const data = await response.json();
+        const data = await response.json();
 
-      if (response.status === 200) {
-        setDeleteBtnText('Delete');
-        // console.log('Announcement deleted successfully!');
-        window.location.reload();
-      } else {
-        // console.log('Failed to delete the announcement!');
-        // console.log(data);
+        if (response.status === 200) {
+          setDeleteBtnText('Delete');
+          window.location.reload();
+        } else {
+          // Handle deletion failure
+        }
+      } catch (error) {
+        // Handle error
       }
-    } catch (error) {
-      // console.error('Error deleting announcement:', error);
     }
   };
 
@@ -54,7 +54,11 @@ const ACard = ({ id, date, description, groupOrSubgroup }) => {
     <div className='card-a'>
       <div className='top-an'>
         <div className='grpclass-a'>{date}</div>
-        {isAdmin && <div className='del-an' onClick={handleDelete}>{deleteBtnText}</div>}
+        {isAdmin && (
+          <div className='del-an' onClick={handleDelete}>
+            {deleteBtnText}
+          </div>
+        )}
       </div>
       <div className='info-a'>
         {showFullDescription ? (
